refactor: remove cjs build from vite core
