--- packages/vite/index.d.cts
@@ -1,6 +1,0 @@
-/**
- * @deprecated The CJS build of Vite's Node API is deprecated. See https://vite.dev/guide/troubleshooting.html#vite-cjs-node-api-deprecated for more details.
- */
-declare const module: any
-
-export = module
